+++ src/index.html
@@ -4,8 +4,15 @@
     <meta charset="UTF-8">
     <title>marching cubes</title>
     <link rel="stylesheet" href="index.css">
+    <script defer src="./functions/triTable.js"></script>
+    <script defer src="./functions/edge.js"></script>
+    <script defer src="./functions/vectorCalculations.js"></script>
+    <script defer src="./functions/noiseFunction.js"></script>
+    <script defer src="./functions/matrix.js"></script>
+    <script defer src="./functions/viewportGL.js"></script>
+    <script defer src="./functions/CubeTriCoords.js"></script>
+    <script defer src="./functions/perlin-noise-3d.min.js"></script>
     <script defer src="getVal.js"></script>
-    <script src="../functions/gl-matrix-min.js" defer></script>
     <style>
       .inp{
         width: 800px;
